requirements.txt
selenium
webdriver-manager
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
supabase
pylint==3.0.3
black==24.1.1 